Reverting pybind plugin change from master branch.
The latest fix commit: f0a98e5cf24fb298b08400e1187c85e88ef79a66 caused
an issue when trying to compile brocade-tunnel.yang. The error is:

FATAL: unmapped type (interface:interface-type)

The "interface:interface-type" definition is part of a union under
the "vrrp:track-ifname".  However, it is not found in the class_map
which it is assumed to be.

With the previous find_definitions/find_child_definitions, the typedef
was found so reverting back to the version.  However, will need to revisit.